chore/#3 : 액티비티, 스크린 분리, 컴포넌트 분리
